# apps/web/openapi/petclinic-api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Petclinic MSS301 — Aggregated API",
    "version": "1.0.0",
    "description": "Aggregated from: auth, customers, vets, visits, workflow, billing, products, reviews"
  },
  "servers": [
    {
      "url": "http://localhost:8180",
      "description": "API Gateway"
    }
  ],
...
  "paths": {
...
    "/api/v1/vets/{vetId}/work-schedule": {
      "get": {
        "tags": ["Vet Work Schedule"],
        "summary": "List work-schedule slots of a vet",
        "description": "Trả empty list nếu vet chưa set lịch. 404 nếu vet không tồn tại. Slot sort theo (workday Mon→Sun, workHour 8h→20h).",
        "operationId": "listVetWorkSchedule",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkScheduleSlotResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Vet Work Schedule"],
        "summary": "Replace work-schedule of a vet (idempotent)",
        "description": "Body: { slots: [{workday, workHour}, ...] }. Empty slots = clear all (tương đương DELETE). Duplicate slot trong request → 400 error.slot-duplicate. Invalid enum value → 400 ProblemDetail. Vet không tồn tại → 404.",
        "operationId": "replaceVetWorkSchedule",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkScheduleSlotResponse"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Vet Work Schedule"],
        "summary": "Clear all work-schedule slots of a vet",
        "operationId": "clearVetWorkSchedule",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/vets/{vetId}/photo": {
      "get": {
        "tags": ["Vet Photo"],
        "summary": "Get photo metadata + presigned URL",
        "description": "Trả 404 nếu vet chưa upload photo. Presigned URL TTL config qua petclinic.files.presigned-ttl (default 1h).",
        "operationId": "getVetPhoto",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VetPhotoResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Vet Photo"],
        "summary": "Upload or replace vet photo (idempotent)",
        "description": "Multipart field 'file'. Max 10MB (config). Content-type: image/jpeg|png|webp. File quá lớn → 400 error.file-too-large. Type khác → 400 error.unsupported-media. Re-upload overwrite cùng object key qua files-service + upsert DB row.",
        "operationId": "uploadVetPhoto",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VetPhotoResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Vet Photo"],
        "summary": "Delete vet photo (xoá object qua files-service trước, DB row sau)",
        "operationId": "deleteVetPhoto",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/vets/me/photo": {
      "get": {
        "tags": ["Vet Me (self-service)"],
        "summary": "Avatar của vet đang login (metadata + presigned URL)",
        "description": "Trả 404 nếu vet chưa upload photo. vetId resolve từ JWT claim.",
        "operationId": "getMyVetPhoto",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VetPhotoResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Vet Me (self-service)"],
        "summary": "Upload/replace avatar của vet đang login (idempotent)",
        "description": "Multipart field 'file'. Max 10MB. Content-type: image/jpeg|png|webp. Mỗi lần đổi ảnh reset status PENDING — chờ STAFF/ADMIN duyệt mới hiển thị public.",
        "operationId": "uploadMyVetPhoto",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VetPhotoResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Vet Me (self-service)"],
        "summary": "Xóa avatar của vet đang login",
        "operationId": "deleteMyVetPhoto",
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/vets/{vetId}/ratings": {
      "get": {
        "tags": ["Vet Ratings"],
        "summary": "List ratings of a vet (paginated)",
        "description": "Vet không tồn tại → 404. Use ?page=0&size=20&sort=rateDate,desc. Optional ?year=YYYY lọc rating theo năm của rateDate.",
        "operationId": "listVetRatings",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageRatingResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Vet Ratings"],
        "summary": "Add or update rating of a vet (UPSERT)",
        "description": "score 1-5 (out of range → 400 ProblemDetail). customerName lấy từ JWT claim 'username' (auth-service ký vào access token) — client không thể giả mạo identity. Vet không tồn tại → 404. JWT thiếu claim 'username' → 400 (token version cũ → cần login lại). POST lần thứ 2 của cùng customer trên cùng vet → UPSERT (update rating cũ + đổi rateDate). Unique (vet_id, customer_name) ở DB layer. Body field customerName (nếu có) bị Jackson silently ignore.",
        "operationId": "addVetRating",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vets/{vetId}/badges": {
      "get": {
        "tags": ["Vet Badges"],
        "summary": "List badges of a vet (paginated)",
        "description": "Vet không tồn tại → 404. Use ?page=0&size=20&sort=awardedDate,desc.",
        "operationId": "listVetBadges",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageBadgeResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Vet Badges"],
        "summary": "Add badge to a vet",
        "description": "title từ enum BadgeTitle (ROOKIE/EXPERIENCED/MASTER/SURGERY_EXPERT/RESEARCH_AWARD/TOP_RATED). awardedDate không được trong tương lai → 400 error.date-future. 1 vet có thể nhận cùng badge nhiều lần (vd kỷ niệm năm). Vet không tồn tại → 404.",
        "operationId": "addVetBadge",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BadgeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadgeResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vets/me": {
      "get": {
        "tags": ["Vet Me (self-service)"],
        "summary": "Profile của vet đang login",
        "operationId": "getMyVetProfile",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VetResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Vet Me (self-service)"],
        "summary": "Sửa profile của vet đang login (PATCH partial)",
        "operationId": "updateMyVetProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VetResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vets/{vetId}/ratings/summary": {
      "get": {
        "tags": ["Vet Ratings"],
        "summary": "Get rating summary of a vet",
        "description": "Trả {count, average (null nếu count=0), distribution Map<1..5, count>}. Distribution luôn có đủ 5 key, value=0 nếu không có rating ở score đó.",
        "operationId": "getVetRatingsSummary",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vets/me/ratings": {
      "get": {
        "tags": ["Vet Me (self-service)"],
        "summary": "Ratings customer chấm cho vet đang login (paginated)",
        "operationId": "listMyRatings",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageRatingResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vets/me/badges": {
      "get": {
        "tags": ["Vet Me (self-service)"],
        "summary": "Badges của vet đang login (paginated)",
        "operationId": "listMyBadges",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageBadgeResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "unpaged": {
            "type": "boolean"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          }
        }
      },
...
      "WorkScheduleRequest": {
        "type": "object",
        "properties": {
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkScheduleSlotResponse"
            }
          }
        },
        "required": ["slots"]
      },
      "WorkScheduleSlotResponse": {
        "type": "object",
        "properties": {
          "workday": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ]
          },
          "workHour": {
            "type": "string",
            "enum": [
              "HOUR_8_9",
              "HOUR_9_10",
              "HOUR_10_11",
              "HOUR_11_12",
              "HOUR_12_13",
              "HOUR_13_14",
              "HOUR_14_15",
              "HOUR_15_16",
              "HOUR_16_17",
              "HOUR_17_18",
              "HOUR_18_19",
              "HOUR_19_20"
            ]
          }
        },
        "required": ["workHour", "workday"]
      },
      "VetPhotoResponse": {
        "type": "object",
        "properties": {
          "vetId": {
            "type": "integer",
            "format": "int64"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          },
          "presignedUrl": {
            "type": "string",
            "format": "url"
          },
          "status": {
            "type": "string"
          },
          "reviewedBy": {
            "type": "string"
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time"
          },
          "rejectReason": {
            "type": "string"
          }
        }
      },
...
      "SpecialtyResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "VetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "cardCode": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "vetBillId": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "resume": {
            "type": "string"
          },
          "specialties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpecialtyResponse"
            }
          },
          "photoUrl": {
            "type": "string"
          },
          "averageRating": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "RatingRequest": {
        "type": "object",
        "properties": {
          "score": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "predefinedDescription": {
            "type": "string",
            "enum": ["POOR", "AVERAGE", "GOOD", "EXCELLENT"]
          }
        },
        "required": ["score"]
      },
      "RatingResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "vetId": {
            "type": "integer",
            "format": "int64"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "predefinedDescription": {
            "type": "string",
            "enum": ["POOR", "AVERAGE", "GOOD", "EXCELLENT"]
          },
          "customerName": {
            "type": "string"
          },
          "rateDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "BadgeRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "enum": [
              "ROOKIE",
              "EXPERIENCED",
              "MASTER",
              "SURGERY_EXPERT",
              "RESEARCH_AWARD",
              "TOP_RATED"
            ]
          },
          "awardedDate": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          }
        },
        "required": ["awardedDate", "title"]
      },
      "BadgeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "vetId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "enum": [
              "ROOKIE",
              "EXPERIENCED",
              "MASTER",
              "SURGERY_EXPERT",
              "RESEARCH_AWARD",
              "TOP_RATED"
            ]
          },
          "awardedDate": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "UpdateVetRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "vetBillId": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "resume": {
            "type": "string"
          },
          "specialtyNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        }
      },
...
      "PageRatingResponse": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RatingResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "RatingSummaryResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "average": {
            "type": "number",
            "format": "double"
          },
          "distribution": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "PageBadgeResponse": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BadgeResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {}
  }
}